Extract Halitosis::ErrorEntry serializer
