+++ spec/halitosis/rails/exception_serializer_spec.rb
@@ -5,77 +5,19 @@ return unless defined?(Rails)
 RSpec.describe Halitosis::ExceptionSerializer, :rails do
   let(:exception) { StandardError.new("Something went wrong") }
 
-  describe "::ErrorEntry" do
-    it "raises MissingOption when error is not provided" do
-      expect { described_class::ErrorEntry.new }.to raise_error(Halitosis::MissingOption)
-    end
-
-    it "renders an empty hash when no fields are defined" do
-      result = described_class::ErrorEntry.new(error: exception).as_json
-
-      expect(result).to eq({})
-    end
-  end
-
-  describe "::ClassMethods" do
-    subject(:entry_class) { Class.new(described_class::ErrorEntry) }
-
-    %i[id code title status detail].each do |field|
-      describe "##{field}" do
-        it "defines a #{field} attribute" do
-          entry_class.public_send(field) { field.to_s.upcase }
-
-          result = entry_class.new(error: exception).as_json
-
-          expect(result[field.to_s]).to eq(field.to_s.upcase)
-        end
-      end
-    end
-
-    describe "#source_pointer" do
-      it "renders source with pointer key" do
-        entry_class.source_pointer { "/data/attributes/title" }
-
-        result = entry_class.new(error: exception).as_json
-
-        expect(result["source"]).to eq("pointer" => "/data/attributes/title")
-      end
-    end
-
-    describe "#source_parameter" do
-      it "renders source with parameter key" do
-        entry_class.source_parameter { "sort" }
-
-        result = entry_class.new(error: exception).as_json
-
-        expect(result["source"]).to eq("parameter" => "sort")
-      end
-    end
-
-    describe "#source_header" do
-      it "renders source with header key" do
-        entry_class.source_header { "Authorization" }
-
-        result = entry_class.new(error: exception).as_json
-
-        expect(result["source"]).to eq("header" => "Authorization")
-      end
-    end
-
-    it "exposes the exception as `error` in attribute blocks" do
-      entry_class.detail { error.message }
-
-      result = entry_class.new(error: exception).as_json
-
-      expect(result["detail"]).to eq("Something went wrong")
-    end
-  end
-
   describe ".build" do
     it "returns an ExceptionSerializer instance" do
-      result = described_class.build(exception)
+      result = described_class.build(exception) { code { "error" } }
 
       expect(result).to be_a(described_class)
+    end
+
+    it "raises ArgumentError when called without a block" do
+      expect { described_class.build(exception) }.to raise_error(ArgumentError, /requires a block/)
+    end
+
+    it "raises ArgumentError when the block defines no fields" do
+      expect { described_class.build(exception) {} }.to raise_error(ArgumentError, /at least one attribute, link, or meta field/)
     end
 
     it "wraps the exception in a single-element errors array" do
@@ -93,12 +35,6 @@ RSpec.describe Halitosis::ExceptionSerializer, :rails do
       error_object = result["errors"].first
       expect(error_object["code"]).to eq("unauthorized")
       expect(error_object["detail"]).to eq("Something went wrong")
-    end
-
-    it "renders an empty error entry when no fields are defined" do
-      result = described_class.build(exception).as_json
-
-      expect(result["errors"]).to eq([{}])
     end
 
     context "with source shortcuts" do
@@ -136,7 +72,7 @@ RSpec.describe Halitosis::ExceptionSerializer, :rails do
     end
 
     it "accepts a custom error_serializer_class" do
-      custom_class = Class.new(described_class::ErrorEntry) do
+      custom_class = Class.new(Halitosis::ErrorEntry) do
         code { "custom" }
       end
